% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleMicrobiomeData.R
\docType{data}
\name{exampleMicrobiomeData}
\alias{exampleMicrobiomeData}
\title{Example microbiome data}
\format{
A data frame with 183 rows (genus) and 80 columns (samples). All values are non-negative integers.
}
\usage{
exampleMicrobiomeData
}
\description{
A data frame contains the count table of the 183 genus in the 80 samples.
}
\keyword{datasets}
