% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molBV.R
\name{molBV}
\alias{molBV}
\title{Perform the calculation of the molBV scores}
\usage{
molBV(inputData, outputTable = NULL)
}
\arguments{
\item{inputData}{Either a biom file or a matrix-like R object with sample IDs as columns and genera as rows.}

\item{outputTable}{The path and name of the output table to be generated. Default is NULL.}
}
\value{
A data frame contains the sample map with the molBV socres.
\itemize{
  \item SID - Sample IDs.
  \item Column 2 to 12 - Counts of the genus that used to calculated the scores.
  \item molBV - Calculated molBV scores.
}
}
\description{
Perform the calculation of the molBV scores
}
\details{
\itemize{
  \item If \code{inputData} is biom file, package 'phyloseq' is needed to process it. Some warnings related to the greengenes database can be ignored.
  \item If \code{inputData} is a matrix/data frame, please make sure that the genera labels (row names) match the green-genes convention, which is necessary when calculating the score for each reference frame.
  \item If \code{outputTable} is specified, a table contains the sample map with the molBV socres will be generated.
}
}
\examples{
map_with_molBVScores <- molBV(exampleMicrobiomeData, outputTable = NULL)
print(map_with_molBVScores)
range(map_with_molBVScores$molBV)

exampleBiomFile <- system.file("extdata", "converted_otu.biom", package = "molBV")
map_with_molBVScores <- molBV(exampleBiomFile, outputTable = "map_with_molBV.txt")
#A txt file "map_with_molBV.txt" will be generated in current working directory.

}
\references{
Usyk et al. The Immune Landscape of Molecular Bacterial Vaginosis and HPV Natural History. https://doi.org/10.21203/rs.3.rs-613095/v1
}
