% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecuL.R
\docType{data}
\name{molecuL}
\alias{molecuL}
\title{The coefficients in the formula}
\format{
A data frame with 10 rows and 6 variables:
\describe{
  \item{bug1}{Lactobacillus}
  \item{bug2}{10 genus used in the formula}
  \item{R2}{R square values}
  \item{beta}{beta coefficients}
  \item{intercept}{intercept coefficients}
  \item{weight}{from 10 to 1}
}
}
\usage{
molecuL
}
\description{
A data frame contains the coefficients for each BV association ratio (column beta and intercept).
}
\keyword{datasets}
